+++ src/generator.rs
@@ -3,8 +3,8 @@ pub trait Generator {
     fn next(&mut self) -> Option<Self::Item>;
 }
 
-impl Generator for f64 {
-    type Item = f64;
+impl Generator for f32 {
+    type Item = f32;
     fn next(&mut self) -> Option<Self::Item> {
         Some(*self)
     }
