+++ src/vco.rs
@@ -1,12 +1,12 @@
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 use crate::generator::Generator;
 
-type WaveShape = fn(f64, f64, f64) -> f32;
+type WaveShape = fn(f32, f32, f32) -> f32;
 
 #[derive(Copy, Clone)]
 pub struct VCO<T, U> {
     shape: WaveShape,
-    time: f64,
+    time: f32,
     pub freq: T,
     pub volume: U
 }
@@ -22,7 +22,7 @@ impl<T, U> VCO<T, U> {
     }
 }
 
-impl<T, U> Generator for VCO<T, U> where T: Generator<Item=f64>, U: Generator<Item=f64>{
+impl<T, U> Generator for VCO<T, U> where T: Generator<Item=f32>, U: Generator<Item=f32>{
     type Item = f32;
     fn next(&mut self) -> Option<Self::Item> {
         self.time += 1. / 44_100.;
